requirements/test.txt
# Dependencies related to testing linkanywhere-back
-r base.txt

# Mocking
mock==2.0.0

# Fixtures
factory-boy==2.8.1

# Test Runner
django-nose==1.4.4
nose-progressive==1.5.1
pytest-django==3.1.2
pytest-sugar==0.8.0
pytest-cov==2.5.1

# Code Coverage
coverage==4.3.4

# Style Enforcement
flake8==3.3.0
